Swift 3 manual migration changes.

+++ WordPress/Classes/ViewRelated/People/PersonViewController.swift
@@ -375,7 +375,7 @@ private extension PersonViewController {
 private extension PersonViewController {
 
     var isMyself : Bool {
-        return blog.account!.userID == person.ID || blog.account!.userID == person.linkedUserID
+        return blog.account!.userID.intValue == person.ID || blog.account!.userID.intValue == person.linkedUserID
     }
 
     var isFullnamePrivate: Bool {
